Refactored and fixed import
